requirements.txt
Django
django-ninja
django-cors-headers
Pillow
pydantic[email]
pyjwt
python-dotenv
pytz